Prevent Raven/TraceKit from taking over window.onerror
TraceKit (and thus, Raven) automatically replaces window.onerror
with its own handler (and chains the old one) when it is set up.
The old workaround for that was to just unsubscribe the setup
script (which loads Raven and sends errors to window.onerror
again once it is loaded). That caused two bugs:
* for all the errors that were triggered before Raven finished
  loading, all global.error handlers other than Raven were
  triggered twice
* under certain timing conditions (probably two errors in very
  quick succession), mw.track could get into a state
  where there was a backlog of items in its trackQueue, in which
  case unsubscribing that handler did not help to prevent infinite
  loops: Raven's error handler called the old onerror handler,
  which pushed the error on top of trackQueue, and once Raven's
  error handler finished, it was called again with the new item.

Instead, just restore the onerror handler after TraceKit's setup
code has run. This has the disadvantage that TraceKit is not
the top-level error handler anymore, so an exception in some
other global.error subscriber can prevent error reporting,
but that's an OK tradeoff for making the whole setup simpler.

Bug: T143279
Change-Id: I4b5b74a257ba98d7078c5b045abbb1e2fbf5185c